+++ src/main/java/com/epam/mvc/smoke/service/impl/UserServiceImpl.java
@@ -53,7 +53,7 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public User findById(long id) {
+    public User findById(int id) {
         return this.userRepository.findById(id);
     }
 
@@ -63,7 +63,7 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public void delete(long id) {
+    public void delete(int id) {
         this.userRepository.delete(id);
     }
 }
